pytorch-example/app/requirements.txt
numpy
Pillow
torch 
torchvision
urllib3